! SQLite only takes integer fields as a primary key. (install_2-0_sqlite.sql) [Bug 1554]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@6799 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/install_2-0_sqlite.sql
@@ -2026,7 +2026,7 @@ INSERT INTO {$db_prefix}spiders	(id_spider, spider_name, user_agent, ip_info) VA
 #
 
 CREATE TABLE {$db_prefix}subscriptions(
-  id_subscribe smallint primary key,
+  id_subscribe integer primary key,
   name varchar(60) NOT NULL,
   description varchar(255) NOT NULL,
   cost text NOT NULL,
